fix metadata structure
