+++ src/vs/workbench/contrib/positronOutputWebview/browser/notebookOutputWebviewService.ts
@@ -6,7 +6,7 @@ import { createDecorator } from 'vs/platform/instantiation/common/instantiation'
 import { IOverlayWebview } from 'vs/workbench/contrib/webview/browser/webview';
 import { ILanguageRuntime, ILanguageRuntimeMessageOutput } from 'vs/workbench/services/languageRuntime/common/languageRuntimeService';
 import { Event } from 'vs/base/common/event';
-import { IPyWidgetHtmlData } from 'vs/workbench/contrib/positronIPyWidgets/browser/positronIPyWidgetsService';
+import { IPyWidgetHtmlData } from 'vs/workbench/services/positronIPyWidgets/common/positronIPyWidgetsService';
 
 export const POSITRON_NOTEBOOK_OUTPUT_WEBVIEW_SERVICE_ID = 'positronNotebookOutputWebview';
