+++ src/vs/workbench/contrib/positronOutputWebview/browser/notebookOutputWebviewServiceImpl.ts
@@ -17,7 +17,7 @@ import { IWebviewService, WebviewInitInfo } from 'vs/workbench/contrib/webview/b
 import { asWebviewUri } from 'vs/workbench/contrib/webview/common/webview';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { ILanguageRuntime, ILanguageRuntimeMessageWebOutput, RuntimeOutputKind } from 'vs/workbench/services/languageRuntime/common/languageRuntimeService';
-import { IPyWidgetHtmlData } from 'vs/workbench/contrib/positronIPyWidgets/browser/positronIPyWidgetsService';
+import { IPyWidgetHtmlData } from 'vs/workbench/services/positronIPyWidgets/common/positronIPyWidgetsService';
 
 export class PositronNotebookOutputWebviewService implements IPositronNotebookOutputWebviewService {
